Converted the HotkeySettings UI into a QDialog. Merged HotkeySettings into HotkeySettingsDialog.
