+++ lib/autokey/qtui/common.py
@@ -16,10 +16,17 @@
 import re
 
 from PyQt4.QtCore import QFile
+from PyQt4.QtGui import QMessageBox
 from PyQt4 import uic
 
 EMPTY_FIELD_REGEX = re.compile(r"^ *$", re.UNICODE)
 
+def validate(expression, message, widget, parent):
+    if not expression:
+        QMessageBox.critical(parent, message, message)
+        if widget is not None:
+            widget.setFocus()
+    return expression
 
 def _get_ui_qfile(name: str):
     """
