+++ lib/autokey/qtui/resources/ui/hotkeysettings.ui
@@ -1,33 +1,36 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>Form</class>
- <widget class="QWidget" name="Form">
+ <class>Dialog</class>
+ <widget class="QDialog" name="Dialog">
+  <property name="windowModality">
+   <enum>Qt::ApplicationModal</enum>
+  </property>
   <property name="geometry">
    <rect>
     <x>0</x>
     <y>0</y>
     <width>546</width>
-    <height>150</height>
+    <height>186</height>
    </rect>
   </property>
   <property name="windowTitle">
-   <string>Form</string>
+   <string>Set Hotkey</string>
   </property>
   <layout class="QGridLayout" name="gridLayout">
-   <item row="1" column="1">
-    <widget class="QPushButton" name="altButton">
+   <item row="1" column="0">
+    <widget class="QPushButton" name="controlButton">
      <property name="text">
-      <string>Alt</string>
+      <string>Control</string>
      </property>
      <property name="checkable">
       <bool>true</bool>
      </property>
     </widget>
    </item>
-   <item row="1" column="3">
-    <widget class="QPushButton" name="superButton">
+   <item row="1" column="1">
+    <widget class="QPushButton" name="altButton">
      <property name="text">
-      <string>Super</string>
+      <string>Alt</string>
      </property>
      <property name="checkable">
       <bool>true</bool>
@@ -44,10 +47,10 @@
      </property>
     </widget>
    </item>
-   <item row="1" column="0">
-    <widget class="QPushButton" name="controlButton">
+   <item row="1" column="3">
+    <widget class="QPushButton" name="superButton">
      <property name="text">
-      <string>Control</string>
+      <string>Super</string>
      </property>
      <property name="checkable">
       <bool>true</bool>
@@ -64,33 +67,6 @@
      </property>
     </widget>
    </item>
-   <item row="2" column="0" colspan="3">
-    <widget class="QLabel" name="keyLabel">
-     <property name="text">
-      <string>Key: %s</string>
-     </property>
-    </widget>
-   </item>
-   <item row="0" column="0">
-    <widget class="QLabel" name="modsLabel">
-     <property name="text">
-      <string>Modifiers:</string>
-     </property>
-    </widget>
-   </item>
-   <item row="3" column="1">
-    <spacer name="verticalSpacer">
-     <property name="orientation">
-      <enum>Qt::Vertical</enum>
-     </property>
-     <property name="sizeHint" stdset="0">
-      <size>
-       <width>20</width>
-       <height>40</height>
-      </size>
-     </property>
-    </spacer>
-   </item>
    <item row="1" column="5">
     <widget class="QPushButton" name="metaButton">
      <property name="text">
@@ -101,6 +77,13 @@
      </property>
     </widget>
    </item>
+   <item row="2" column="0" colspan="3">
+    <widget class="QLabel" name="keyLabel">
+     <property name="text">
+      <string>Key: %s</string>
+     </property>
+    </widget>
+   </item>
    <item row="2" column="3" colspan="3">
     <widget class="QPushButton" name="setButton">
      <property name="text">
@@ -108,26 +91,80 @@
      </property>
     </widget>
    </item>
+   <item row="3" column="1">
+    <spacer name="verticalSpacer">
+     <property name="orientation">
+      <enum>Qt::Vertical</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>81</width>
+       <height>27</height>
+      </size>
+     </property>
+    </spacer>
+   </item>
    <item row="4" column="0" colspan="6">
-    <widget class="Line" name="line">
+    <widget class="QDialogButtonBox" name="buttonBox">
      <property name="orientation">
       <enum>Qt::Horizontal</enum>
+     </property>
+     <property name="standardButtons">
+      <set>QDialogButtonBox::Cancel|QDialogButtonBox::Ok</set>
+     </property>
+    </widget>
+   </item>
+   <item row="0" column="0">
+    <widget class="QLabel" name="modsLabel">
+     <property name="text">
+      <string>Modifiers:</string>
      </property>
     </widget>
    </item>
   </layout>
-  <zorder>modsLabel</zorder>
-  <zorder>controlButton</zorder>
-  <zorder>altButton</zorder>
-  <zorder>shiftButton</zorder>
-  <zorder>superButton</zorder>
-  <zorder>hyperButton</zorder>
-  <zorder>setButton</zorder>
-  <zorder>keyLabel</zorder>
-  <zorder>line</zorder>
-  <zorder>verticalSpacer</zorder>
-  <zorder>metaButton</zorder>
  </widget>
+ <tabstops>
+  <tabstop>controlButton</tabstop>
+  <tabstop>altButton</tabstop>
+  <tabstop>shiftButton</tabstop>
+  <tabstop>superButton</tabstop>
+  <tabstop>hyperButton</tabstop>
+  <tabstop>metaButton</tabstop>
+  <tabstop>setButton</tabstop>
+ </tabstops>
  <resources/>
- <connections/>
+ <connections>
+  <connection>
+   <sender>buttonBox</sender>
+   <signal>accepted()</signal>
+   <receiver>Dialog</receiver>
+   <slot>accept()</slot>
+   <hints>
+    <hint type="sourcelabel">
+     <x>224</x>
+     <y>167</y>
+    </hint>
+    <hint type="destinationlabel">
+     <x>160</x>
+     <y>132</y>
+    </hint>
+   </hints>
+  </connection>
+  <connection>
+   <sender>buttonBox</sender>
+   <signal>rejected()</signal>
+   <receiver>Dialog</receiver>
+   <slot>reject()</slot>
+   <hints>
+    <hint type="sourcelabel">
+     <x>292</x>
+     <y>167</y>
+    </hint>
+    <hint type="destinationlabel">
+     <x>286</x>
+     <y>124</y>
+    </hint>
+   </hints>
+  </connection>
+ </connections>
 </ui>
